+++ .gitignore
@@ -1,6 +1,7 @@
 /target
 /Cargo.lock
 *.swp
+*~
 # Compiled files
 *.o
 *.so
@@ -12,3 +13,5 @@
 
 # Generated by Cargo
 /target/
+
+
